chores(otel): split logger and pino adapter
